story/pyproject.toml
[project]
name = "story"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "gjc18", email = "gregjc@protonmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "argparse>=1.4.0",
    "lmstudio>=1.5.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
story = "story.cli:cli_main"

[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"
